pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "econ-comm-analysis"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Juan Camilo Reyes" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
dependencies = [
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
]
requires-python = "~=3.10.0"

[tool.flit.module]
name = "argumentation_mining"

[tool.ruff]
line-length = 80
src = ["src"]
include = ["pyproject.toml", "src/**/*.py"]

[tool.ruff.lint]
select = ["ALL"]  # Add import sorting
ignore= ["D203", "D212", "COM812", "TD003"]

[tool.ruff.lint.isort]
known-first-party = ["src"]
force-sort-within-sections = true

[dependency-groups]
dev = [
    "ruff>=0.13.3",
]

argumentation_mining = [
    "openai>=2.1.0",
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.3",
]

article_processing_pipeline = [
    "pandas>=2.3.3",
    "numpy>=2.2.5",
    "sentence-transformers>=3.3.1",
    "datasketch>=1.6.4",
    "gliner>=0.2.0",
    "torch>=2.0.0",
    "scikit-learn>=1.6.0",
]

professional_profiler = [
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
    "pyyaml>=6.0.3",
    "requests>=2.32.3",
    "beautifulsoup4>=4.12.0",
    "nltk>=3.9.1",
    "pydantic-ai>=0.0.14",
]

proquest_scraping = [
    "pandas>=2.3.3",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.3",
    "selenium>=4.0.0",
]

